chore: build grammar
