+++ js-enumerable.code-workspace
@@ -67,7 +67,7 @@
 			}
 		},
 		"editor.codeActionsOnSave": {
-			"source.fixAll.eslint": false
+			"source.fixAll.eslint": "never"
 		},
 		"eslint.validate": [
 			"typescript",
