+++ src/linq/peg/expressionvisitor.ts
@@ -1,4 +1,4 @@
-﻿import { ODataParser, JavascriptParser, transform, parse } from './parser'
+﻿import { transform, parse } from './parser'
 
 import { ExpressionStack } from './expressionstack'
