+++ src/linq/peg/parser/index.ts
@@ -1,5 +1,1 @@
-import ODataParser from './odata-parser'
-import JavascriptParser from './javascript-parser'
-
-export { ODataParser, JavascriptParser }
 export { transform, parse } from './parser'
